+++ libs/qt-style/release_notes.md
@@ -1,5 +1,10 @@
 # ftrack Framework QT Style library release Notes
 
+## v3.0.0
+2024-09-19
+
+* [changed] Update ftrack utils dependency to 3.0.0.
+
 ## v2.2.1
 2024-05-02